Changed int to float and voila…

+++ FasettArduino/ES_Transceiver_07/ES_Transceiver_07.ino
@@ -1,4 +1,4 @@
-int a, b, al, bl;
+float a, b, al, bl;
 
 void setup(){
   Serial.begin(115200);
